データ定義をmodelsパッケージに切り分け
